requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-login==0.6.3
flask-wtf==1.1.1
flask-cors==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
werkzeug==2.3.7
email-validator==2.0.0
bcrypt==4.0.1
pyjwt==2.8.0
marshmallow==3.20.1
